move rx utils in separate file
